directly pass residue data into ContinuousEuler to support serialization and avoid path issues when reloading; maintain back compatibility
